+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.5.0] 2021-05-31
+### Added
+- Add command-line arguments for coinjoin filtering and address prefixes
+### Changed
+- Adapted TagPack schema
+- Changed schema of address/cluster relation tables
+- Speed-up tests
+
 ## [0.4.5] - 2020-11-19
 ### Added
 - Added Dockerfile for `spark-submit` job